# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Population Health Data Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Domain {
  id: number;
  slug: string;
  name: string;
  indicatorCount: number;
  description: string;
}

export interface Indicator {
  id: number;
  name: string;
  domainId: number;
  domainName: string;
  domainSlug: string;
  state: string;
  county: string;
  region: string;
  value: string;
  year: string;
  source: string;
  sourceUrl: string;
  /** 1-5 authority ranking (5 = most authoritative) */
  authorityRank: number;
  notes?: string;
}

export type CountyState = (typeof CountyState)[keyof typeof CountyState];

export const CountyState = {
  TN: "TN",
  VA: "VA",
} as const;

export interface County {
  name: string;
  state: CountyState;
  region: string;
  fips: string;
}

export type SearchRequestState =
  (typeof SearchRequestState)[keyof typeof SearchRequestState];

export const SearchRequestState = {
  TN: "TN",
  VA: "VA",
} as const;

export interface SearchRequest {
  query: string;
  state?: SearchRequestState;
  counties?: string[];
  domain?: string;
}

export interface SearchResponse {
  query: string;
  interpretation: string;
  results: Indicator[];
  totalCount: number;
}

export interface SavedSource {
  id: number;
  indicatorId: number;
  indicatorName: string;
  county: string;
  state: string;
  value: string;
  year: string;
  source: string;
  sourceUrl: string;
  domainName: string;
  savedAt: string;
  notes?: string;
}

export interface SaveSourceRequest {
  indicatorId: number;
  notes?: string;
}

export type PortalSummaryStateBreakdown = {
  TN: number;
  VA: number;
};

export interface PortalSummary {
  totalDomains: number;
  totalIndicators: number;
  totalCounties: number;
  totalSaved: number;
  stateBreakdown: PortalSummaryStateBreakdown;
  recentlyViewed: Indicator[];
}

export type ListIndicatorsParams = {
  /**
   * Filter by domain slug
   */
  domain?: string;
  /**
   * Filter by state
   */
  state?: ListIndicatorsState;
  /**
   * Filter by county name
   */
  county?: string;
};

export type ListIndicatorsState =
  (typeof ListIndicatorsState)[keyof typeof ListIndicatorsState];

export const ListIndicatorsState = {
  TN: "TN",
  VA: "VA",
} as const;

export type ListCountiesParams = {
  /**
   * Filter by state
   */
  state?: ListCountiesState;
  /**
   * Filter by region (e.g. Northeast Tennessee, Southwest Virginia)
   */
  region?: string;
};

export type ListCountiesState =
  (typeof ListCountiesState)[keyof typeof ListCountiesState];

export const ListCountiesState = {
  TN: "TN",
  VA: "VA",
} as const;