# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Population Health Data Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id: number;
  slug: string;
  name: string;
  indicatorCount: number;
  description: string;
}