# lib/api-zod/src/generated/types/indicator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Population Health Data Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Indicator {
  id: number;
  name: string;
  domainId: number;
  domainName: string;
  domainSlug: string;
  state: string;
  county: string;
  region: string;
  value: string;
  year: string;
  source: string;
  sourceUrl: string;
  /** 1-5 authority ranking (5 = most authoritative) */
  authorityRank: number;
  notes?: string;
}