# lib/api-zod/src/generated/types/portalSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Population Health Data Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Indicator } from "./indicator";
import type { PortalSummaryStateBreakdown } from "./portalSummaryStateBreakdown";

export interface PortalSummary {
  totalDomains: number;
  totalIndicators: number;
  totalCounties: number;
  totalSaved: number;
  stateBreakdown: PortalSummaryStateBreakdown;
  recentlyViewed: Indicator[];
}