# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Population Health Data Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchRequestState } from "./searchRequestState";

export interface SearchRequest {
  query: string;
  state?: SearchRequestState;
  counties?: string[];
  domain?: string;
}