# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Population Health Data Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Indicator } from "./indicator";

export interface SearchResponse {
  query: string;
  interpretation: string;
  results: Indicator[];
  totalCount: number;
}